Implement #9 DefaultViewModel
- ItemViewModel의 상위 추상층 추가
