+++ app/src/main/java/com/example/bikini_android/ui/feeds/adapter/FeedItemGridViewModel.kt
@@ -1,20 +1,11 @@
 package com.example.bikini_android.ui.feeds.adapter
 
-import androidx.databinding.Bindable
-import com.example.bikini_android.BR
 import com.example.bikini_android.R
 
 /**
  * @author bsgreentea
  */
-class FeedItemGridViewModel :FeedItemViewModel() {
-
-    @get: Bindable
-    var imageUri = ""
-        set(value) {
-            field = value
-            notifyPropertyChanged(BR.imageUri)
-        }
+class FeedItemGridViewModel : FeedItemViewModel() {
 
     override fun getLayoutRes(): Int = R.layout.item_feed_grid
 }
